Move IntArray2DJsonConverter to Serialization namespace
